Added AIAction, reworked towers
Added AIAction
- Has the Actions (build, destroy, queue) as booleans
- has the x and y coordinate
- and has a supporting parameter for whether you choose queue or build
- reworked build tower to use this
Reworked Entity
- added attack_damage
- added value
Reworked towers
- Tower now has three build-in functions; built, update, & tower activation
- built called when the tower is first placed down
- update called whenever world updates
- and tower activation either shoots in set range or generates money
Our Tower classes didn't have the values needed to separate them, I also game each class an update function so whenever we need to update a class we can just go to the class itself instead of looking in a long script file (it's not that long in this case, but in a larger codebase this would be horrendous).
I also refactored update towers to reflect this change.
Towers don't really need to read the entire gamestate, I just gave them the entity grid for now so since they need to see the entities to shoot them, house has the entire gamestate because player 1 and 2 money is in there, if we moved them to say Playerbase, I'd pass playerbase instead.
Also cooldown didn't work as intended, current implementation would just have the tower activate even if nothing is nearby then go back to dormitory

Also I gave each tower subclass their separate variables that differentiate them though looking at the rules it seems we don't have intial features. I'll keep them for now and add a TODO for when the base game is finished