Game runs
I fixed some bugs that were occurring to omitted lines of code and typos. The game can now do a full turn sequence with all of the phases.

Update Demons still needs to be made and seeing as we didn't git revert my changes we need to discuss the map making thing